==> notebooks/02_mysql_connection.ipynb <==
# Connection to mysql

import mysql.connector

conn = mysql.connector.connect(
    host="127.0.0.1",
    port=3307,
    user="sadop_user",
    password="1234"
)

cursor = conn.cursor()
print("Connected as sadop_user")




# 1.Creation de la data base 


cursor.execute("USE SADOP_BDD")
print("Database selected")

## 1.2Users Tabel

cursor.execute("""
CREATE TABLE user (
    user_id INT PRIMARY KEY,
    full_name VARCHAR(100),
    email VARCHAR(120),
    country CHAR(2),
    signup_date DATE
)
""")

print("✅ users table created")


## 1.3 Accounts table

cursor.execute("""
CREATE TABLE accounts (
    account_id INT PRIMARY KEY,
    user_id INT,
    account_type ENUM('SAVINGS','CHECKING','BUSINESS'),
    balance DECIMAL(12,2),
    created_at DATE,
    FOREIGN KEY (user_id) REFERENCES user(user_id)
)
""")

print("✅ accounts table created")


## 1.4 Transactions table 

cursor.execute("""
CREATE TABLE transactions (
    transaction_id INT PRIMARY KEY,
    account_id INT,
    amount DECIMAL(10,2),
    transaction_type ENUM('DEBIT','CREDIT'),
    transaction_date DATE,
    status ENUM('SUCCESS','FAILED'),
    FOREIGN KEY (account_id) REFERENCES accounts(account_id)
)
""")

print("✅ transactions table created")


## 1.5 log table

cursor.execute("""
CREATE TABLE logs (
    log_id INT PRIMARY KEY,
    user_id INT,
    log_level ENUM('INFO','WARN','ERROR'),
    message TEXT,
    created_at DATETIME,
    FOREIGN KEY (user_id) REFERENCES user(user_id)
)
""")

print("✅ logs table created")


## 1.6 Commit & Verify

cursor.execute("SHOW TABLES")
tables = cursor.fetchall()
print("📦 Tables in database:")
for t in tables:
    print("-", t[0])


## 1.7 instalation of faker to insert data 

pip install Faker

for table in ["user", "accounts", "transactions", "logs"]:
    print(f"\n📄 {table.upper()} STRUCTURE")
    cursor.execute(f"DESCRIBE {table}")
    for row in cursor.fetchall():
        print(row)
