==> notebooks/04_load_csv_into_mysql.ipynb <==
# 4 load_csv_into_mysql

## 4.1 Export Full MySQL Database Backup


import pandas as pd
import mysql.connector
from mysql.connector import Error
import numpy as np
from faker import Faker
import random
import mysql.connector

## 4.2 MySQL Connection Configuration


conn = mysql.connector.connect(
    host="127.0.0.1",
    port=3307,
    user="sadop_user",
    password="1234",
    database="SADOP_BDD"
)
cursor = conn.cursor()
print("Connected to SADOP database")

## 4.4 Drop Existing Tables

We drop tables in reverse dependency order.


users_df = pd.read_csv("../data/users.csv")
accounts_df = pd.read_csv("../data/Accounts.csv")
transactions_df = pd.read_csv("../data/transactions.csv")
logs_df = pd.read_csv("../data/logs.csv")


user_query = """
INSERT INTO user (user_id, full_name, email, country, signup_date)
VALUES (%s, %s, %s, %s, %s)
"""

cursor.executemany(user_query, users_df.values.tolist())
conn.commit()

print(f"👤 Inserted {len(users_df)} users")


account_query = """
INSERT INTO accounts (account_id, user_id, account_type, balance, created_at)
VALUES (%s, %s, %s, %s, %s)
"""

cursor.executemany(account_query, accounts_df.values.tolist())
conn.commit()

print(f"🏦 Inserted {len(accounts_df)} accounts")

transaction_query = """
INSERT INTO transactions (transaction_id, account_id, amount, transaction_type, transaction_date,status)
VALUES (%s, %s, %s, %s, %s, %s)
"""

cursor.executemany(transaction_query, transactions_df.values.tolist())
conn.commit()

print(f"💸 Inserted {len(transactions_df)} transactions")


log_query = """
INSERT INTO logs (log_id, user_id, log_level, message, created_at)
VALUES (%s, %s, %s, %s, %s)
"""

cursor.executemany(log_query, logs_df.values.tolist())
conn.commit()

print(f"📜 Inserted {len(logs_df)} logs")


for table in ["user", "accounts", "transactions", "logs"]:
    cursor.execute(f"SELECT COUNT(*) FROM {table}")
    count = cursor.fetchone()[0]
    print(f"{table}: {count} rows")


cursor.close()
conn.close()
print("🔒 MySQL connection closed")
