==> notebooks/06_MySQL_Monitoring_Setup.ipynb <==
# 6. MySQL Monitoring Setup 


import mysql.connector
from faker import Faker
import random
from datetime import datetime, timedelta
import pandas as pd
print("All set up")

## 6.1 Connect to MySQL


conn = mysql.connector.connect(
    host="127.0.0.1",
    port=3307,
    user="sadop_user",
    password="1234",
    database="sadop_DB"
)
cursor = conn.cursor()
print("Connected to SADOP database")

## 6.2 Check Current Slow Query Log Status


cursor.execute("SHOW VARIABLES LIKE 'slow_query_log';")
print("slow_query_log:", cursor.fetchall())

cursor.execute("SHOW VARIABLES LIKE 'slow_query_log_file';")
print("slow_query_log_file:", cursor.fetchall())

cursor.execute("SHOW VARIABLES LIKE 'long_query_time';")
print("long_query_time:", cursor.fetchall())


## 6.3 Enable Slow Query Log


cursor.execute("SET GLOBAL slow_query_log = 'ON';")
print("Slow query log enabled")


## 6.4 Set Slow Query Threshold


# Set the slow query threshold to 0.5 seconds
cursor.execute("SET GLOBAL long_query_time = 0.4;")

# Verify immediately
cursor.execute("SHOW VARIABLES LIKE 'long_query_time';")
result = cursor.fetchall()
print("long_query_time:", result)


## 6.5 Verify Configuration


cursor.execute("SHOW VARIABLES LIKE 'slow_query_log';")
print("slow_query_log:", cursor.fetchall())

cursor.execute("SHOW VARIABLES LIKE 'long_query_time';")
print("long_query_time:", cursor.fetchall())

cursor.execute("SHOW VARIABLES LIKE 'slow_query_log_file';")
print("slow_query_log_file:", cursor.fetchall())


cursor.execute("SELECT @@GLOBAL.long_query_time;")
print("Global long_query_time:", cursor.fetchone())
