==> notebooks/07_Capture Performance Metrics.ipynb <==
# 7. Capture Performance Metrics

## 7.1 Parse Slow Query Log


import os
import pandas as pd

# Path to slow query log
slow_log_path = r"C:\ProgramData\MySQL\MySQL Server 8.0\Data\DESKTOP-GK9TB81-slow.log"

# Check if file exists
if not os.path.exists(slow_log_path):
    print("Slow query log file not found!")
else:
    with open(slow_log_path, "r") as f:
        lines = f.readlines()
    print(f"Loaded {len(lines)} lines from slow query log")


## 7.2 Extract Query Execution Time and SQL


# with this we can stor the query time and the query text that are slower then the 0.5s

import re

queries = []
times = []

for i, line in enumerate(lines):
    match = re.search(r'# Query_time: ([0-9.]+)', line)
    if match:
        times.append(float(match.group(1)))
        # Take next line(s) as SQL (skip comments starting with #)
        sql_line = ""
        j = i + 1
        while j < len(lines) and not lines[j].startswith("#"):
            sql_line += lines[j].strip() + " "
            j += 1
        queries.append(sql_line.strip())



## 7.3 Save Metrics to CSV for ML/RL

# her we save them in a CSV file 

df_metrics = pd.DataFrame({
    "query_time": times,
    "query": queries
})

# Save CSV
metrics_csv_path = r"C:\Users\pc\data science\SADOP\data\slow_query_metrics.csv"
df_metrics.to_csv(metrics_csv_path, index=False)
print(f"Metrics saved to {metrics_csv_path}")
